delete asv1 related code in the sdk
