+++ packages/sdk/base/src/phoneNumbers.test.ts
@@ -1,5 +1,5 @@
 import { soliditySha3 } from 'web3-utils'
-import { getPhoneHash, isE164Number } from './phoneNumbers'
+import { isE164Number } from './phoneNumbers'
 const sha3 = (v: string) => soliditySha3({ type: 'string', value: v })
 
 const TEST_PHONE_NUMBERS = {
@@ -30,27 +30,6 @@ const TEST_PHONE_NUMBERS = {
 }
 
 describe('Phone number formatting and utilities', () => {
-  describe('Phone hashing', () => {
-    it('Hashes an valid number without a salt', () => {
-      expect(getPhoneHash(sha3, TEST_PHONE_NUMBERS.VALID_E164, '')).toBe(
-        '0x483128504c69591aed5751690805ba9aad6c390644421dc189f6dbb6e085aadf'
-      )
-    })
-    it('Hashes an valid number with a salt', () => {
-      expect(getPhoneHash(sha3, TEST_PHONE_NUMBERS.VALID_E164, 'abcdefg')).toBe(
-        '0xf08257f6b126597dbd090fecf4f5106cfb59c98ef997644cef16f9349464810c'
-      )
-    })
-    it('Throws for an invalid number', () => {
-      try {
-        getPhoneHash(sha3, TEST_PHONE_NUMBERS.VALID_US_1, '')
-        fail('expected an error')
-      } catch (error) {
-        // Error expected
-      }
-    })
-  })
-
   describe('Other phone helper methods', () => {
     it('checks if number is e164', () => {
       // @ts-ignore
