+++ packages/sdk/phone-utils/src/index.ts
@@ -1,7 +1,6 @@
 export * from './countries'
 export { Countries, LocalizedCountry } from './countries'
 export * from './getCountryEmoji'
-export * from './getPhoneHash'
 export * from './inputValidation'
 export * from './io'
 export {
